# optimize performance : Hessian2Input.readObject() for 'B' or 'b'