--- a/src/test/java/com/caucho/hessian/io/Hessian2InputTest.java
+++ b/src/test/java/com/caucho/hessian/io/Hessian2InputTest.java
@@ -1,3 +1,19 @@
+/*
+ * Licensed to the Apache Software Foundation (ASF) under one or more
+ * contributor license agreements.  See the NOTICE file distributed with
+ * this work for additional information regarding copyright ownership.
+ * The ASF licenses this file to You under the Apache License, Version 2.0
+ * (the "License"); you may not use this file except in compliance with
+ * the License.  You may obtain a copy of the License at
+ *
+ *     http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
 package com.caucho.hessian.io;
 
 import org.junit.Assert;
@@ -11,11 +27,11 @@ public class Hessian2InputTest {
 
     SerializerFactory serializerFactory = new SerializerFactory();
 
-    int[] sizeList = new int[] {
-            0, 2, 16, 512, 1024,
-            0x8000, 0x8000 + 1,
-            64 * 1024, 64 * 1024 * 1024
-    };
+    int[]             sizeList          = new int[] {
+                                        0, 2, 16, 512, 1024,
+                                        0x8000, 0x8000 + 1,
+                                        64 * 1024, 64 * 1024 * 1024
+                                        };
 
     @Test
     public void testSerialize() {